travel_bot_backend/requirements.txt
fastapi
uvicorn
openai
langchain
python-dotenv
requests